+++ Server API/Controllers/locationsController.cs
@@ -85,6 +85,11 @@ namespace Server_API.Controllers
             if (id != Location.id)
                 return BadRequest("PUT URL and ID in the location do not match");
 
+            // Verify token
+            string msg = AuthorizeHeader.VerifyTokenWithUserId(ActionContext, Location.user_id);
+            if (!String.IsNullOrEmpty(msg))
+                return BadRequest(msg);
+
             // Convert the Location_API to the EntityModel location
             location loc = ConvertLocationApiToLocation(Location);
 
@@ -100,6 +105,11 @@ namespace Server_API.Controllers
         {
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
+
+            // Verify token
+            string msg = AuthorizeHeader.VerifyTokenWithUserId(ActionContext, Location.user_id);
+            if (!String.IsNullOrEmpty(msg))
+                return BadRequest(msg);
 
             // Convert the Location_API to the EntityModel location
             location loc = ConvertLocationApiToLocation(Location);
@@ -122,6 +132,11 @@ namespace Server_API.Controllers
             {
                 return NotFound();
             }
+
+            // Verify token
+            string msg = AuthorizeHeader.VerifyTokenWithUserId(ActionContext, location.user_id);
+            if (!String.IsNullOrEmpty(msg))
+                return BadRequest(msg);
 
             db.locations.Remove(location);
             await db.SaveChangesAsync();
